old_requirements.txt
absl-py==2.0.0
accelerate==0.24.1
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
altair==5.1.2
anyio==3.7.1
apex @ file:///mnt/build/apex
appdirs==1.4.4
argcomplete==3.1.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.29.5
azure-identity==1.15.0
Babel==2.13.1
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
boilerpy3==1.0.7
cachetools==5.3.2
cattrs==23.2.1
Cerberus==1.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.7
comm==0.2.0
construct==2.5.3
contourpy==1.1.1
coverage==6.3.1
cryptography==41.0.5
cycler==0.12.1
databricks-cli==0.18.0
debugpy==1.8.0
decorator==5.1.1
deeplog==0.0.3
deepspeed==0.11.1
defusedxml==0.7.1
distro==1.8.0
docker==6.1.3
docopt==0.6.2
elastic-transport==7.16.0
elasticsearch==7.17.9
entrypoints==0.4
Events==0.5
exceptiongroup==1.1.3
executing==2.0.1
fairscale==0.4.13
farm-haystack==1.22.1
fastjsonschema==2.19.0
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.44.3
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
hjson==3.1.0
horovod==0.24.2
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
igraph==0.11.2
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflect==7.0.0
iniconfig==2.0.0
intel-openmp==2021.4.0
ipykernel==6.26.0
ipython==8.12.3
isoduration==20.11.0
itsdangerous==2.1.2
jarowinkler==1.2.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.2
kiwisolver==1.4.5
lazy-imports==0.3.1
lightning-utilities==0.9.0
lit==17.0.4
logpai==1.0.0
logparser3==1.0.4
lxml==4.9.3
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mkl-include==2021.4.0
mlflow==2.8.1
monotonic==1.6
more-itertools==10.1.0
mpi4py==3.1.1
mpmath==1.3.0
msal==1.24.1
msal-extensions==1.0.0
msccl @ git+https://github.com/microsoft/msccl-tools.git@030a750fc56e03fb863ffe48268672d84d4cdc7b
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nebulaml==0.16.6
nest-asyncio==1.5.8
networkx==3.1
ninja==1.10.2
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
num2words==0.5.13
numpy==1.22.2
oauthlib==3.2.2
onnx==1.15.0
onnxruntime-training==1.16.0+cu118
openai==1.3.3
opencensus==0.11.3
opencensus-context==0.1.3
opencensus-ext-azure==1.1.11
opencensus-ext-logging==0.1.1
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
plotly==5.18.0
pluggy==1.3.0
portalocker==2.8.2
posthog==3.0.2
prometheus-client==0.18.0
prompt-toolkit==3.0.41
prompthub-py==4.0.0
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.2
pytest-mpi==0.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-ptrace==0.9.8
pytorch-lightning==1.9.3
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
quantulum3==0.9.0
querystring-parser==1.2.4
rank-bm25==0.2.2
rapidfuzz==2.7.0
referencing==0.31.0
regex==2022.3.2
requests==2.31.0
requests-cache==0.9.8
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.13.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
seqeval==1.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
sseclient-py==1.8.0
stack-data==0.6.3
streamlit==1.28.2
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tbb==2021.10.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
terminado==0.18.0
texttable==1.7.0
threadpoolctl==3.2.0
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch @ file:///mnt/build/training/install/internal_packages/download_internal_packages/torch_package/torch-2.0.0-cp38-cp38-linux_x86_64.whl#sha256=0639f76c2f9d69fdd0b21e54e3c9519d4b861c15350ce265e7bc58281f05b3b7
torch-nebula==0.16.6
torch-ort==1.16.0
torch-train==0.0.3
torchaudio==2.0.1+cu117
torchmetrics==0.11.3
torchsnapshot==0.1.0
torchtrain==0.4.13
torchvision==0.15.1+cu117
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.34.1
triton==2.0.0
tutel @ https://github.com/microsoft/tutel/archive/refs/tags/v0.1.5.tar.gz#sha256=f9fd4a0be6da14dfcc9dbfcaeb9b32aa027b3aebe44a80209db8ae0ecd212947
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
url-normalize==1.4.3
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
yarl==1.9.2
z3-solver==4.12.2.0
zipp==3.17.0
